Cleaning up the code a bit -- moving the util into the hook and adding comments
