+++ packages/dashboard/src/components/Input/MetadataFilter/index.tsx
@@ -1,3 +1,2 @@
 export { default as MetadataFilterEditor } from './Editor';
 export { type Condition, type MetadataFilterType, conditions, type MetadataFilterGroupType, type Combination, combinations } from './types';
-export { filterContactsByMetadata } from './util';
